requirements.txt
alembic==0.9.7
annoy==1.11.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
nltk==3.2.5
olefile==0.44
pathlib2==2.3.0
Pillow==4.3.0
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.2
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
